Extract statements_params.go
