+++ statements_params.go
@@ -1,0 +1,62 @@
+package main
+
+import (
+	"context"
+	"maps"
+	"slices"
+
+	"github.com/apstndb/lox"
+	"github.com/cloudspannerecosystem/memefish"
+	"github.com/cloudspannerecosystem/memefish/ast"
+	"github.com/samber/lo"
+	"spheric.cloud/xiter"
+)
+
+type ShowParamsStatement struct{}
+
+func (s *ShowParamsStatement) Execute(ctx context.Context, session *Session) (*Result, error) {
+	strMap := make(map[string]string)
+	for k, v := range session.systemVariables.Params {
+		strMap[k] = v.SQL()
+	}
+
+	rows := slices.SortedFunc(
+		xiter.MapLower(maps.All(session.systemVariables.Params), func(k string, v ast.Node) Row {
+			return toRow(k, lo.Ternary(lox.InstanceOf[ast.Type](v), "TYPE", "VALUE"), v.SQL())
+		}),
+		ToSortFunc(func(r Row) string { return r.Columns[0] }))
+
+	return &Result{
+		ColumnNames:   []string{"Param_Name", "Param_Kind", "Param_Value"},
+		Rows:          rows,
+		KeepVariables: true,
+	}, nil
+}
+
+type SetParamTypeStatement struct {
+	Name string
+	Type string
+}
+
+func (s *SetParamTypeStatement) Execute(ctx context.Context, session *Session) (*Result, error) {
+	if expr, err := memefish.ParseType("", s.Type); err != nil {
+		return nil, err
+	} else {
+		session.systemVariables.Params[s.Name] = expr
+		return &Result{KeepVariables: true}, nil
+	}
+}
+
+type SetParamValueStatement struct {
+	Name  string
+	Value string
+}
+
+func (s *SetParamValueStatement) Execute(ctx context.Context, session *Session) (*Result, error) {
+	if expr, err := memefish.ParseExpr("", s.Value); err != nil {
+		return nil, err
+	} else {
+		session.systemVariables.Params[s.Name] = expr
+		return &Result{KeepVariables: true}, nil
+	}
+}
